adds demo-data; updates readme and moves scripts to python path (#1)
