--- v4l2_camera/v4l2_camera.yaml
@@ -1,32 +1,0 @@
-%YAML 1.2
-# SPDX-FileCopyrightText: Copyright (c) 2022-2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
----
-source:  # V4L2VideoCaptureOp
-  # | Input device.  On devkit, /dev/video0 is generally HDMI in
-  device: "/dev/video1"
-
-  # | App will auto-select default width and height if not provided
-  # | For performance, you may want to specify these parameters yourself.
-  # | See this app's readme file for details.
-  width: 352 #320 #160
-  height: 288 #240 #120
-
-  # | App will auto-select the default "pixel_format" for your device if not provided.
-  # | See this app's readme file for details.
-  # pixel_format: "AR24"
-  pixel_format: "YUYV"
-
-visualizer:  # Holoviz
